feat(fsp plugin): implement vdi creation by relying on sr.scan()
This commit replaces current workaround by using a subdirectory
named directories in the file-uri directory that is passed during
sr.create(). User has to create a symlink in this directory and
then issue a sr.scan() to trigger the creation of a new vdi.
To remove a volume, user has to remove the symlink and issue a
sr.scan().

Signed-off-by: Matias Ezequiel Vara Larsen <matias.vara@vates.fr>