Add support for destroy volumes and snapshots
This commit improves how the driver destroys volumes and snapshots. The
commit creates a tree structure in which snapshots are the parent of volumes
and other snapshots. The root of this tree is the oldest snapshot. The button
of the three is the latest snapshot. Snapshots can be destroy from top to
button (from oldest to lastest) or from button to top (from latest to oldest).
Those snapshots that have volumes cannot be destroyed until all volumes are
destroyed.

Signed-off-by: Matias Ezequiel Vara Larsen <matias.vara@vates.fr>