Use "sr-{uuid}" instead of label to name the zpool 2/2
This part deals with the volume manipulation.  Since the volume plugin
does not use a "SR metadata" struct we have to introduce a separate
helper function here - and since we mostly care about it when
accessing a volume, simply handle creation of volume name in one
place.

This is unfortunately needed in other loosely-related code units, so
hardcoding goes on...